+++ iosApp/iosApp/Views/Dashboard/AccountView.swift
@@ -1,0 +1,15 @@
+//
+//  AccountView.swift
+//  iosApp
+//
+//  Created by Nasir Ahmed Momin on 07/06/22.
+//  Copyright © 2022 Mutual Mobile. All rights reserved.
+//
+
+import SwiftUI
+
+struct AccountView: View {
+    var body: some View {
+        Text("Account View")
+    }
+}
