+++ iosApp/iosApp/Views/Dashboard/ReportSheetView.swift
@@ -1,0 +1,15 @@
+//
+//  ReportSheetView.swift
+//  iosApp
+//
+//  Created by Nasir Ahmed Momin on 07/06/22.
+//  Copyright © 2022 Mutual Mobile. All rights reserved.
+//
+
+import SwiftUI
+
+struct ReportSheetView: View {
+    var body: some View {
+        Text("Report View")
+    }
+}
